Fix issues with non-visible global variables.
When using dplyr and other functions that have an implicit environment, we see frequent notes for "No visible binding for global variable" for each of the used variables.  By importing `.data` from rlang, we make it explicit that variables are defined in the `.data` frame using the syntax `.data$score` for example.  See https://www.r-bloggers.com/no-visible-binding-for-global-variable/